backend/pyproject.toml
[tool.poetry]
name = "pdf-processor-backend"
version = "0.1.0"
description = "PDF processing backend (Flask)"
authors = ["jenifer <jenifer@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-cors = "^4.0.0"
pypdf = "^3.15.0"
Pillow = "^10.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
